requirements.txt
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.0.11
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
locust==2.23.1
MarkupSafe==2.1.5
msgpack==1.0.7
prometheus_client==0.20.0
psutil==5.9.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pyzmq==25.1.2
redis==5.0.1
requests==2.31.0
roundrobin==0.0.4
setuptools==69.1.1
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.2
Faker~=23.2.1